Refactoring for improved MVC architecture

+++ demo/src/main/java/com/dooapp/fxform/MyBean.java
@@ -3,8 +3,6 @@ package com.dooapp.fxform;
 import com.dooapp.fxform.annotation.FormFactory;
 import javafx.beans.property.*;
 import org.hibernate.validator.constraints.Email;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * Created at 26/09/12 14:23.<br>
@@ -30,7 +28,7 @@ public class MyBean {
 
     private final ObjectProperty<Subject> subject = new SimpleObjectProperty<Subject>();
 
-    private final DoubleProperty age = new SimpleDoubleProperty();
+    private final IntegerProperty age = new SimpleIntegerProperty();
 
     protected MyBean(String name, String email, String message, String website, boolean subscribe, Subject subject) {
         this.name.set(name);
